Fix router.on() parameter validation
I noticed that one of the test cases in `errors.test.js` would fail when tested with `ignoreTrailingSlash=true`.

This adds a bunch of error tests for when `ignoreTrailingSlash` is `true` and fixes the `router.on()` code so that all the tests pass. This also speeds up validation a bit since the same parameters won't be validated multiple times when the input `method` is an array.