Always pass derived constraints to .find
We were already allocating the object every time anyways, and it's going to be necessary for almost every request, because almost every request will have a Host header. Let's just be explicit and pass it always. An alternative would be to check if the object was passed and instantiate it otherwise, but I think that makes for a nicer API and worse performance which is the wrong tradeoff.