Always derive constraint values for custom constraints using mustMatchWhenDerived
For the special constraints which should match requests that have a value for the constraint only with routes that have a constraint set, we need to ensure we always derive a value for the constraint even if no routes are using that constraint yet. That way, we know if the route has a value that must match. This was a bug caused by the performance optimization of trying to avoid deriving constraint values for constraints that aren't in use -- we must consider all custom mustMatchWhenDerived constraints in use in order to check if the value is present.

See #196